requirements.txt
streamlit>=1.49.0
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
requests>=2.31.0
llama-index>=0.14.0
llama-index-embeddings-fastembed>=0.5.0
fastembed>=0.7.0
watchdog>=6.0.0
python-frontmatter>=1.1.0